+++ src/si/gto76/funphototime/filterthreads/ThresholdingThread.java
@@ -10,9 +10,9 @@ public class ThresholdingThread extends FilterThread {
 	
 	private int value;
 	
-	public ThresholdingThread( BufferedImage imgIn, BufferedImage imgOut, int value, MyInternalFrame selectedFrame ) {
+	public ThresholdingThread( BufferedImage imgIn, BufferedImage imgOut, double value, MyInternalFrame selectedFrame ) {
 		super(imgIn, imgOut, selectedFrame);
-		this.value = value;
+		this.value = (int) value;
 	}
 		
 	protected int filtriraj(int rgb) {
